+++ src/main/java/com/teamproject/trackers/biz/product/ProductDetailVO.java
@@ -1,14 +1,14 @@
 package com.teamproject.trackers.biz.product;
 
 public class ProductDetailVO {
-	private int p_id;
+	private long p_id;
 	private String img;
 	
 	
-	public int getP_id() {
+	public long getP_id() {
 		return p_id;
 	}
-	public void setP_id(int p_id) {
+	public void setP_id(long p_id) {
 		this.p_id = p_id;
 	}
 	public String getImg() {
